+++ .github/actions/rust-test-action/action.yml
@@ -19,7 +19,7 @@ runs:
     # Pull cargo nextest binaries
     - uses: taiki-e/install-action@v2
       with:
-        tool: nextest
+        tool: nextest,cargo-hack
 
     - name: Check formatting
       working-directory: ${{ inputs.working-directory }}
@@ -29,6 +29,11 @@ runs:
     - name: Check clippy
       working-directory: ${{ inputs.working-directory }}
       run: cargo clippy --all-features --all-targets --target-dir ${{ inputs.target-folder }} -- -Dwarnings ${{ inputs.additional-flags }}
+      shell: bash
+
+    - name: Check clippy (each features)
+      working-directory: ${{ inputs.working-directory }}
+      run: cargo hack clippy --each-feature --all-targets --target-dir ${{ inputs.target-folder }} -- -Dwarnings ${{ inputs.additional-flags }}
       shell: bash
 
     - name: Run all tests
