+++ dcipher-agents/src/ser.rs
@@ -1,9 +1,6 @@
 //! Helper traits used to serialize and deserialize various types into EVM bytes.
 
-use crate::agents::blocklock::contracts::TypesLib;
-use crate::ibe_helper::IbeIdentityOnBn254G1Ciphertext;
 use alloy::primitives::Bytes;
-use alloy::sol_types::SolType;
 use ark_ec::AffineRepr;
 
 /// Serialize into an EVM Bytes type.
@@ -36,56 +33,66 @@ impl EvmSerialize for ark_ec::short_weierstrass::Affine<ark_bn254::g1::Config> {
     }
 }
 
-#[derive(thiserror::Error, Debug)]
-pub enum IbeIdentityOnBn254G1CiphertextError {
-    #[error("abi decode error")]
-    AbiDecode(#[from] alloy::sol_types::Error),
+#[cfg(feature = "blocklock")]
+pub use blocklock::*;
 
-    #[error("invalid ephemeral pk")]
-    InvalidEphemeralPk,
-}
+#[cfg(feature = "blocklock")]
+mod blocklock {
+    use super::*;
+    use crate::agents::blocklock::contracts::TypesLib;
+    use crate::ibe_helper::IbeIdentityOnBn254G1Ciphertext;
+    use alloy::sol_types::SolType;
 
-impl EvmDeserialize for IbeIdentityOnBn254G1Ciphertext {
-    type Error = IbeIdentityOnBn254G1CiphertextError;
+    #[derive(thiserror::Error, Debug)]
+    pub enum IbeIdentityOnBn254G1CiphertextError {
+        #[error("abi decode error")]
+        AbiDecode(#[from] alloy::sol_types::Error),
 
-    fn deser(bytes: &Bytes) -> Result<Self, Self::Error> {
-        use ark_ff::PrimeField;
+        #[error("invalid ephemeral pk")]
+        InvalidEphemeralPk,
+    }
 
-        let ciphertext = TypesLib::Ciphertext::abi_decode(bytes)?;
-        let x0: [u8; 32] = ciphertext.u.x[0].to_be_bytes();
-        let x1: [u8; 32] = ciphertext.u.x[1].to_be_bytes();
-        let y0: [u8; 32] = ciphertext.u.y[0].to_be_bytes();
-        let y1: [u8; 32] = ciphertext.u.y[1].to_be_bytes();
+    impl EvmDeserialize for IbeIdentityOnBn254G1Ciphertext {
+        type Error = IbeIdentityOnBn254G1CiphertextError;
 
-        let x0 = ark_bn254::Fq::from_be_bytes_mod_order(&x0);
-        let x1 = ark_bn254::Fq::from_be_bytes_mod_order(&x1);
-        let y0 = ark_bn254::Fq::from_be_bytes_mod_order(&y0);
-        let y1 = ark_bn254::Fq::from_be_bytes_mod_order(&y1);
+        fn deser(bytes: &Bytes) -> Result<Self, Self::Error> {
+            use ark_ff::PrimeField;
 
-        let x = ark_bn254::Fq2::new(x0, x1);
-        let y = ark_bn254::Fq2::new(y0, y1);
+            let ciphertext = TypesLib::Ciphertext::abi_decode(bytes)?;
+            let x0: [u8; 32] = ciphertext.u.x[0].to_be_bytes();
+            let x1: [u8; 32] = ciphertext.u.x[1].to_be_bytes();
+            let y0: [u8; 32] = ciphertext.u.y[0].to_be_bytes();
+            let y1: [u8; 32] = ciphertext.u.y[1].to_be_bytes();
 
-        // Use unchecked to return results instead of panicking
-        let eph_pk = ark_ec::short_weierstrass::Affine::new_unchecked(x, y);
-        if !eph_pk.is_on_curve()
-            || !eph_pk.is_in_correct_subgroup_assuming_on_curve()
-            || eph_pk.is_zero()
-        {
-            Err(Self::Error::InvalidEphemeralPk)?
+            let x0 = ark_bn254::Fq::from_be_bytes_mod_order(&x0);
+            let x1 = ark_bn254::Fq::from_be_bytes_mod_order(&x1);
+            let y0 = ark_bn254::Fq::from_be_bytes_mod_order(&y0);
+            let y1 = ark_bn254::Fq::from_be_bytes_mod_order(&y1);
+
+            let x = ark_bn254::Fq2::new(x0, x1);
+            let y = ark_bn254::Fq2::new(y0, y1);
+
+            // Use unchecked to return results instead of panicking
+            let eph_pk = ark_ec::short_weierstrass::Affine::new_unchecked(x, y);
+            if !eph_pk.is_on_curve()
+                || !eph_pk.is_in_correct_subgroup_assuming_on_curve()
+                || eph_pk.is_zero()
+            {
+                Err(Self::Error::InvalidEphemeralPk)?
+            }
+
+            Ok(Self::new(eph_pk))
         }
-
-        Ok(Self::new(eph_pk))
     }
 }
 
 #[cfg(test)]
 pub(crate) mod tests {
-    use super::*;
-
+    #[cfg(all(feature = "blocklock", feature = "ibe"))] // uses blocklock types & ibe
     pub(crate) mod bn254 {
-        use super::*;
-        use crate::agents::blocklock::contracts::BLS;
-        use crate::ibe_helper::IbeCiphertext;
+        use super::super::*;
+        use crate::agents::blocklock::contracts::{BLS, TypesLib};
+        use crate::ibe_helper::{IbeCiphertext, IbeIdentityOnBn254G1Ciphertext};
         use alloy::primitives::U256;
         use alloy::sol_types::SolValue;
         use ark_ff::{BigInteger, Fp, PrimeField};
