+++ dsigner/Dockerfile
@@ -1,32 +1,5 @@
 # Image must be built from the root
 
-# Image building solidity contracts
-ARG NODE_VERSION=22.3
-FROM europe-west1-docker.pkg.dev/randamu-prod/candyland/node:${NODE_VERSION} AS sol_builder
-WORKDIR /app
-
-RUN curl -L https://foundry.paradigm.xyz | bash
-ENV PATH="/root/.foundry/bin:${PATH}"
-RUN foundryup
-
-FROM sol_builder AS sol_blocklock
-WORKDIR /app/blocklock-solidity
-COPY ./blocklock-solidity/package.json ./
-COPY ./blocklock-solidity/package-lock.json ./
-RUN npm install
-
-COPY ./blocklock-solidity ./
-RUN FOUNDRY_PROFILE=build forge install --no-git && FOUNDRY_PROFILE=build forge build
-
-FROM sol_builder AS sol_randomness
-WORKDIR /app/randomness-solidity
-COPY ./randomness-solidity/package.json ./
-COPY ./randomness-solidity/package-lock.json ./
-RUN npm install
-
-COPY ./randomness-solidity ./
-RUN FOUNDRY_PROFILE=build forge install --no-git && FOUNDRY_PROFILE=build forge build
-
 # Base image for rust
 FROM lukemathwalker/cargo-chef:latest-rust-1 AS chef
 WORKDIR /app
@@ -42,8 +15,6 @@ COPY --from=planner /app/recipe.json recipe.json
 RUN cargo chef cook --release --recipe-path recipe.json
 
 # Build application
-COPY --from=sol_blocklock /app/blocklock-solidity/out ./blocklock-solidity/out
-COPY --from=sol_randomness /app/randomness-solidity/out ./randomness-solidity/out
 COPY . .
 RUN cargo build --release -p dsigner
