+++ randomness-agent/Dockerfile
@@ -9,15 +9,6 @@ RUN curl -L https://foundry.paradigm.xyz | bash
 ENV PATH="/root/.foundry/bin:${PATH}"
 RUN foundryup
 
-FROM sol_builder AS sol_blocklock
-WORKDIR /app/blocklock-solidity
-COPY ./blocklock-solidity/package.json ./
-COPY ./blocklock-solidity/package-lock.json ./
-RUN npm install
-
-COPY ./blocklock-solidity ./
-RUN FOUNDRY_PROFILE=build forge install --no-git && FOUNDRY_PROFILE=build forge build
-
 FROM sol_builder AS sol_randomness
 WORKDIR /app/randomness-solidity
 COPY ./randomness-solidity/package.json ./
@@ -42,7 +33,6 @@ COPY --from=planner /app/recipe.json recipe.json
 RUN cargo chef cook --release --recipe-path recipe.json
 
 # Build application
-COPY --from=sol_blocklock /app/blocklock-solidity/out ./blocklock-solidity/out
 COPY --from=sol_randomness /app/randomness-solidity/out ./randomness-solidity/out
 COPY . .
 RUN cargo build --release -p randomness-agent --example randomness
